Introduce AuthenticationController
This new controller provides functionality to retrieve the current signed in user and also sign in user anonymously, and can be called from anywhere. This decouples authentication from a firestore-only use case.

The implementation relies on an AuthenticationListener to switch out the real and fake authentication controllers for testing purposes.

This commit has not included the Auth module in app module test files(Fragment/Acticity tests), hence these will fail to build with a dagger error. This is fixed in the subsequent commit.