Fix part of #4391: Fix failing build on m1 macs (#5111)
<!-- READ ME FIRST: Please fill in the explanation section below and
check off every point from the Essential Checklist! -->
Fix part of #4391 : Fix failing build on M1 Macs
## Explanation
<!--
- Explain what your PR does. If this PR fixes an existing bug, please
include
- "Fixes #bugnum:" in the explanation so that GitHub can auto-close the
issue
  - when this PR is merged.
  -->
Some M1 Macs fail to build the Oppia app because of unused imports in
some of the proto files in the project. This PR attempts to fix this
issue by removing unused imports in proto files. Files affected by this
fix are;
- profile.proto
- deprecation.proto

This is a temporary fix and more investigation will need to be done to
properly understand why this behavior is erratic and why it only happens
in some instances and not in others. More checks are also needed to
ensure that proto files only have imports that are used in the proto
file before they are merged.

Besides build fixes, this PR also;
- Includes a fix for the `buildifier_download.sh` which previously did
not work for MacOS users.
- Introduces a `static_checks.sh` file to allow contributors to run all
static checks locally before pushing the code to Git.
- Adds a wiki entry in the Static-Analysis-Checks wiki. The entry
contains instructions on how to run static checks locally with the
static_checks.sh file.

## Essential Checklist
<!-- Please tick the relevant boxes by putting an "x" in them. -->
- [x] The PR title and explanation each start with "Fix #bugnum: " (If
this PR fixes part of an issue, prefix the title with "Fix part of
#bugnum: ...".)
- [x] Any changes to
[scripts/assets](https://github.com/oppia/oppia-android/tree/develop/scripts/assets)
files have their rationale included in the PR explanation.
- [x] The PR follows the [style
guide](https://github.com/oppia/oppia-android/wiki/Coding-style-guide).
- [x] The PR does not contain any unnecessary code changes from Android
Studio
([reference](https://github.com/oppia/oppia-android/wiki/Guidance-on-submitting-a-PR#undo-unnecessary-changes)).
- [x] The PR is made from a branch that's **not** called "develop" and
is up-to-date with "develop".
- [x] The PR is **assigned** to the appropriate reviewers
([reference](https://github.com/oppia/oppia-android/wiki/Guidance-on-submitting-a-PR#clarification-regarding-assignees-and-reviewers-section)).

---------

Co-authored-by: Kenneth Murerwa <kkmurerwa@Kenneths-MacBook-Air.local>
Co-authored-by: Adhiambo Peres <59600948+adhiamboperes@users.noreply.github.com>