Fix #5246: "Update the Android FAQ" (#5247)
<!-- READ ME FIRST: Please fill in the explanation section below and
check off every point from the Essential Checklist! -->
## Explanation
<!--
- Explain what your PR does. If this PR fixes an existing bug, please
include
- "Fixes #bugnum:" in the explanation so that GitHub can auto-close the
issue
  - when this PR is merged.
  -->
Fixes #5246: "Update the Android FAQ"

1. Changed the order in which the FAQ questions where arranged by
refactoring the property name of each string to the appropriate one
based on requirements from #5246.
2. Added new questions based on requirements from #5246.
3. Refactored the faq question property names in the tests written in
FAQListFragmentTest.kt according to the new order in which the questions
where arranged.
4. Created a string extension file in string extensions to check if
string resource contains a placeholder, to be able to dynamically
replace the placeholders with the appropriate string values.
5. Used the string extension to check if any string resource file from
the FAQ questions and answers contains a placeholder, then call function
to replace placeholder with appropriate string value.



https://github.com/oppia/oppia-android/assets/54560535/d2c0b6cb-251b-4591-9221-1d95a308cc74


https://github.com/oppia/oppia-android/assets/54560535/7b409e30-e9f6-4b4b-80ed-5449678a184f


## Essential Checklist
<!-- Please tick the relevant boxes by putting an "x" in them. -->
- [x] The PR title and explanation each start with "Fix #bugnum: " (If
this PR fixes part of an issue, prefix the title with "Fix part of
#bugnum: ...".)
- [x] Any changes to
[scripts/assets](https://github.com/oppia/oppia-android/tree/develop/scripts/assets)
files have their rationale included in the PR explanation.
- [x] The PR follows the [style
guide](https://github.com/oppia/oppia-android/wiki/Coding-style-guide).
- [x] The PR does not contain any unnecessary code changes from Android
Studio
([reference](https://github.com/oppia/oppia-android/wiki/Guidance-on-submitting-a-PR#undo-unnecessary-changes)).
- [x] The PR is made from a branch that's **not** called "develop" and
is up-to-date with "develop".
- [x] The PR is **assigned** to the appropriate reviewers
([reference](https://github.com/oppia/oppia-android/wiki/Guidance-on-submitting-a-PR#clarification-regarding-assignees-and-reviewers-section)).