Fix API 21-23 (incl.) crash.
The main problem was the use of computeIfAbsent() which is a Java 8 API
that does NOT desugar during building, and Java 8 APIs are only
available on SDK 24+.

The actual fix introduced a lot of complexity when working with locale
profiles since the utilization of memoized locale profiles now requires
using a coroutine to ensure thread safety (i.e. by leveraging a blocking
dispatcher).

This introduced some complexity in the process for creating locales
since a livelock was possible when trying to use the blocking dispatcher
for the system locale (since a blocking dispatcher is used farther up in
its call chain).

Due to these changes, a refactor of AndroidLocaleProfile has been
included to largely simplify a lot of the previous complexity around the
different states the class could take (these are now explicitly defined
as separate subclasses to the now sealed class AndroidLocaleProfile).

DataProviders.transformAsync() has also been updated to handle caught
exceptions in much the same way as transform(). This helped during
debugging the original crash, so it seems like a good idea to include.