Update Maven dep list checker & updater scripts.
Updates the utilities for the Maven dep list & updater scripts to have
much more robustness when trying to verify the correct repository URLs
for artifact coordinates, and when downloading artifact POM files. The
main robustness changes come in the form of retries and smarter
parallelization and error checking.

These changes were necessary due to one of my workstations frequently
flaking out on large numbers of URL blocks during verification, and this
helps considerably.