Fix #2536: Introduce a central utility for matching item in RecyclerView (#5391)
<!-- READ ME FIRST: Please fill in the explanation section below and
check off every point from the Essential Checklist! -->
## Explanation
Fix #2536

I worked on the below functions renaming the title and adding an
argument for the recyclerviewId to each function to make it reusable, I
also added kdocs to help understand the functions.
P.S:- I created the functions in the RecyclerviewMatcher class.
 
![Screenshot 2024-04-30 at 01 24
45](https://github.com/oppia/oppia-android/assets/54560535/6a0464e0-e72f-4516-83e1-3a53d2025a6b)

![Screenshot 2024-04-30 at 01 24
26](https://github.com/oppia/oppia-android/assets/54560535/34332586-e082-4991-9b6b-b8f8c647e7d0)

![Screenshot 2024-04-30 at 01 24
57](https://github.com/oppia/oppia-android/assets/54560535/7090b959-fb45-4787-9957-7e182d47a042)


## Essential Checklist
<!-- Please tick the relevant boxes by putting an "x" in them. -->
- [x] The PR title and explanation each start with "Fix #bugnum: " (If
this PR fixes part of an issue, prefix the title with "Fix part of
#bugnum: ...".)
- [x] Any changes to
[scripts/assets](https://github.com/oppia/oppia-android/tree/develop/scripts/assets)
files have their rationale included in the PR explanation.
- [x] The PR follows the [style
guide](https://github.com/oppia/oppia-android/wiki/Coding-style-guide).
- [x] The PR does not contain any unnecessary code changes from Android
Studio
([reference](https://github.com/oppia/oppia-android/wiki/Guidance-on-submitting-a-PR#undo-unnecessary-changes)).
- [x] The PR is made from a branch that's **not** called "develop" and
is up-to-date with "develop".
- [x] The PR is **assigned** to the appropriate reviewers
([reference](https://github.com/oppia/oppia-android/wiki/Guidance-on-submitting-a-PR#clarification-regarding-assignees-and-reviewers-section)).