Undo Robolectric upgrade.
Robolectric upgrade is hard: it requires upgrading Espresso in order to
work around https://github.com/robolectric/robolectric/issues/7367 and
https://github.com/robolectric/robolectric/issues/6593. However, newer
Espresso makes use of new manifest features that require upgrading
Bazel: https://github.com/bazelbuild/bazel/issues/14285. Because
upgrading Kotlin is a precursor to the Bazel upgrade, a workaround for
the single test that required the Robolectric upgrade (due to using a
deprecated API from SDK 33 without being able to actually use SDK 33 in
the test) is the best approach here until Robolectric can actually be
upgraded in the future.