Revert "Fix #5357: Remove CDATA from translatable strings." (#5403)
Reverts oppia/oppia-android#5361

There are some formatting issues that cause incorrectness in the HTML
structure, so @translatewiki cannot parse them correctly. This is the
cause of empty translatiion strings in #5274. Mainly, some elements are
missing their closing tags, e.g.:

```xml 
<string name="faq_answer_bug_reporting">&lt;p&gt;&lt;ol&gt;&lt;li&gt;From your %s app home screen, tap the menu in the top left corner.&lt;/li&gt;&lt;li&gt;Tap &lt;strong&gt;Share feedback&lt;/strong&gt;.&lt;/li&gt;&lt;li&gt;Follow the instructions to report the bug or share feedback.&lt;/li&gt;&lt;/p&gt;</string>
```
is missing an \</ol\> tag before the closing \</p\> tag.

I have looked at several of the other "broken" strings, and they have
similar issues. When @translatewiki encounters an unclosed tag, it
continues reading until it encounters a corresponding closing tag, in a
subsequent string, meaning any strings between the two points are
assumed to be part of a single string. This is obviously a serious bug
in translatewiki itself.

I think it is best to revert this PR and fix the issues first before
merging it again.

Co-authored-by: Ben Henning <bhenning@google.com>